# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthenticationController_signUp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignUpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully"
          }
        },
        "summary": "Register a new user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthenticationController_signIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully"
          }
        },
        "summary": "Login a user",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/refresh-tokens": {
      "post": {
        "operationId": "AuthenticationController_refreshTokens",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens refreshed successfully"
          }
        },
        "summary": "Refresh authentication tokens",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/auth/impersonate": {
      "post": {
        "operationId": "AuthenticationController_impersonate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImpersonateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Impersonate a user (Admins only)",
        "tags": [
          "Authentication"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthenticationController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset email sent"
          }
        },
        "summary": "Request a password reset email",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthenticationController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          }
        },
        "summary": "Reset user password",
        "tags": [
          "Authentication"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stansbury Swim API",
    "description": "API documentation for Stansbury Swim backend",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "SignUpDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "firstName": {
            "type": "string",
            "example": "John",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "example": "Doe",
            "description": "Last name"
          },
          "phoneNumber": {
            "type": "string",
            "example": "(555) 555-5555",
            "description": "User phone number"
          },
          "turnstileToken": {
            "type": "string",
            "description": "Cloudflare Turnstile token. Required once TURNSTILE_SECRET_KEY is set."
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "phoneNumber"
        ]
      },
      "SignInDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "refresh-token-value",
            "description": "Refresh token"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ImpersonateDto": {
        "type": "object",
        "properties": {}
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email address"
          },
          "turnstileToken": {
            "type": "string",
            "description": "Cloudflare Turnstile token. Required once TURNSTILE_SECRET_KEY is set."
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "reset-token-value",
            "description": "Password reset token"
          },
          "password": {
            "type": "string",
            "example": "newPassword123",
            "description": "New password"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
...
    }
  }
}